requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdf2image==1.17.0
pytesseract==0.3.10
pypdf==3.17.1
langchain==0.0.267
openai==0.28.0
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
numpy==1.26.3
sqlalchemy==2.0.25
tiktoken==0.5.2 